requirements.txt
transformers
OpenHowNet
pytorch-crf
tensorboard
truecase